Request Azure authorization token with TokenAudience
Azure/go-autorest#258 introduced a new TokenAudience field on the
azure.Environment struct, but it doesn't seem to have made it into this
repository. It's inteded usage is for private and hybrid cloud models of Azure
to set a different audience than the resource manager endpoint. For the existing
cloud environments defined in Azure/go-autorest ResourceManagerEndpoint is
identical to TokenAudience, so this change shouldn't affect existing usage and
only opens for the posibility of setting a different TokenAudience for private
and hybrid cloud models.

This change has been verified to enable authentication with Azure Stack Hub.