Remove the need to specify services for which we have no available versions (#2217)
When a new service is released but provides no usable resources we
currently have to pin to an empty spec to unblock nightly upgrades. This
also meant that once the new service gets usable resources we won't
start picking them up because there's an entry already defined, so it
won't cause an error.

This will still block nightly upgrades, but only when there's an actual
version to fill in.