Change the way we populate function descriptions (#2276)
This PR changes the way we build the description for functions
(invokes). By default, it takes the description of the corresponding
HTTP method (GET or POST). If that description is empty, it falls back
to the old way (use the description of the schema type of the response
object).

This method seems to yield better results on average. Some description
won't be perfect, but they are often framed in terms of "getting things"
rather than "things", which is close to what we need.

The example in #2264 now says "Retrieves information about a gallery
image version." instead of "Specifies information about the gallery
image version that you want to create or update"

Fix https://github.com/pulumi/pulumi-azure-native/issues/2264