Apply case-spacing-insensitive diffing to all location properties (#2281)
All `location` properties in Azure are case- and space- insensitive, so
both "West Europe" and "westeurope" are valid values representing the
same region.

This PR adjusts our diff calculation logic with this special rule. We
get a diff calculated by vanilla platform diffing, then we check if
there is an Update to the property called `location`, and if so, it
normalizes both values and compares those. Any non-significant changes
are ignored.

Fix #1809
Fix #1005