Remove arm2pulumi coverage workflow
The result is not consulted, and workflow failures are ignored. So: just
don't do it.

This commit removes the associated files s3-upload-script.sh and
results.json; the first was a helper for the workflow, and it's unclear
why the second was checked in (perhaps as a basis for comparison? But it
was never updated.)

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>