Remove versions removed in most recent release
Adding these causes nodejs to be unreleasable so we have to remove these. This is unlikely to cause any user impact as they were only released for users of dotnet, python, java and go - if they updated within the last 24 hours and only if they start using the specified new explicit, undocumented versions.

Going to attempt to keep signification updates from:
- AppConfiguration: 2023-03-01
- Communication: 2023-03-31
- Consumption: 2023-03-01