Request JSON output from az for reliable parsing (#2380)
Fixes #2373.

When calling `az`, we would assume its output was JSON. That is the
default, but users can change that. We should explicitly request JSON to
be sure we can parse the result.

Tested locally by setting my default to YAML (`az config set
core.output=yaml`) and verifying that before this change a `pulumi up`
invocation fails, with the change it succeeds.