Improve Refresh accuracy for dictionary and array values (#2378)
Fix #2356

Azure Native has a peculiar way to update inputs when a `pulumi refresh`
calls the provider's Read method. It's impossible to reads "inputs" from
Azure, so we use the following algorithm to refresh existing inputs:
1. Read the current properties of a resource, it's "new outputs".
2. Diff "old outputs" from the state with these new outputs from Azure.
3. See which properties changed/added/removed.
4. Apply only those changed properties to existing inputs from state to
produce new inputs.

This is intended to minimize the noise and diffs in `refresh` and
subsequent `up` commands, which #2356 illustrates.

The problem is that the step (4) is currently operating on top-level
properties only. For map- or array-properties that means that any change
inside them will lead to us copying the entire content of that value to
new inputs. This is the cause for #2356.

This PR changes the step (4) "apply diff" to be a recursive comparison
across maps and arrays. We compare property-by-property and
element-by-element and only copy the granular changes that we identify.

I extended our existing test to cover those scenarios. I also tested
https://github.com/pulumi/pulumi-azure-native/issues/2356#issuecomment-1503672473
and got a nice diff on the last step of Ringo's repro:

<img width="828" alt="image"
src="https://user-images.githubusercontent.com/1454008/231751192-bb9c8656-8404-476c-b95c-56fa492b1bc5.png">

cc @ringods if you want to take a look or give this branch a try locally
(`make provider` then copy `./bin/pulumi-resource-azure-native`
somewhere on your path).