Output v2-removed-resources.yaml
This will become read-only at the point of releasing v2, but can be manually edited if required to re-introduce resources.
This is quite similar in purpose to the v2-removed.json, but that is currently used for removing whole versions of services rather than specific resources. This is not currently used for v2 - the new resource-level removals is probably more powerful and could supersede the v2-remove.json