Identify resources with different paths in v1 and v2
We only check for changes in normalized paths.

We don't warn on path changes when previous version is explicitly excluded.
When we've determined that the v1 version of a resource should not be
considered for the default namespace anymore, we don't need to warn.

Curate MySQL and PostgreSQL to avoid mixing single and flexible server.