Improve autorest 'Service returned an error. Status=<nil> <nil>' message by adding details (#2435)
fixes #2433

When an Azure request made via Autorest fails, the user often (always?)
only sees the unhelpful error message

```
    error: autorest/azure: Service returned an error. Status=<nil> <nil>
```

For the linked issue, for instance, this PR changes it to

```
    error: autorest/azure: Service returned an error. Status=<nil> <nil>. There was a conflict. AlwaysOn cannot be set for this site as the plan does not allow it. For more information on pricing and features, please see: https://aka.ms/appservicepricingdetails
```

The `<nil> <nil>` comes from go-autorest and seems like a bug in that
package. I considered dropping the outer error altogether but there may
be cases where it's more informative.