Update eventhub namespace default resource state for v2
When deleting, we PUT the resource back to a default state. The newer version of the API fails with the following message:

This update of networkRuleSets, sets zero IPRule and VirtualNetworkRules, and the DefaultAction is set to Deny. This would render the namespace resources inaccessible for data-plane requests. If the intention was to restrict data-plane requests to private links only, please set the PublicNetworkAccess property to Disabled. Else either set the DefaultAction property to Allow OR add some allowed IPRules or VirtualNetworkRules in the update request