Add combined docs schema to the dist (#2488)
Enable breaking the dependency of the docs publishing process on
checking out the schema that's being committed in the repository. There
is a filesize limit of 100mb which the full schema current exceeds and
therefore can't be committed which is why we currently have a second
smaller "docs" schema which is comitted but is missing all explicit
versions.

- All files in `dist/` are picked up in the GH release as an asset.
- The explode + impode process includes the whole of the current
committed docs schema plus all explicit versions (without examples).