Disambiguate resources with different paths in v1 and v2 (#2484)
We only check for changes in normalized paths.

Any changes found are printed but no automated action is taken. The output looks like

```
[V1->V2 path change] PrivateEndpointConnection: /subscriptions/{}/resourcegroups/{}/providers/microsoft.cache/redis...
    .../{}/privateendpointconnections/{}
    ...enterprise/{}/privateendpointconnections/{}
```

We don't warn on path changes when previous version is explicitly excluded.

When we've determined that the v1 version of a resource should not be considered for the default namespace anymore, we don't need to warn.