Handle string sets during preview (#2524)
Closes https://github.com/pulumi/pulumi-azure-native/issues/2525

At preview time, we generally just return the inputs back as outputs -
as a best-guess as to what might be available after deployment. However,
for string sets the input and output SDK shapes are different. Therefore
we need to handle this scenario specifically in the "input to output
converter". For now, we return a top-level computed object - which is
unknown at preview.

It might be possible to make this known if all elements in the input
array are also known but this would be significantly more complex with
likely no real-world benefit as the properties within the values still
won't be known and they could just extract a variable with their input
ids as an array which would be easier to loop over if required rather
than extracting the keys from the map.

- Fix preview conversion of StringSet input just gets an unknown output
at preview-time rather than returning the input list.
- Add special test case for this type.
- Fix metadata so we have it as the object type in metadata, but an
array in the schema.

Metadata diff (when converted to yaml using `cat
bin/metadata-compact.json | yj -jy > bin/metadata-compact.yaml`):

```diff
2623,2625c2623
<         type: array
<         items:
<           type: string
---
>         type: object
2657,2659c2655
<         type: array
<         items:
<           type: string
---
>         type: object
7959,7961c7955
<         type: array
<         items:
<           type: string
---
>         type: object
9195,9197c9189
<         type: array
<         items:
<           type: string
---
>         type: object
10645,10647c10637
<         type: array
<         items:
<           type: string
---
>         type: object
12272,12274c12262
<         type: array
<         items:
<           type: string
---
>         type: object
14000,14002c13988
<         type: array
<         items:
<           type: string
---
... snipped ...
```