Exclude resources and invokes within gen.PulumiSchema
Consider whether to add the resource or invoke at the point of generating then adding to the schema & metadata.
- Trying to modify the original set of specs is problematic because the names aren't final until variants & enums have been processed.
- Trying to modify the generated schema is tricky as any associated types may still be left in the schema.

Use a "Constraints" interface to keep the additions to the main schema gen process minimal. This allows us to elegantly express the process for considering whether to add a resource or invoke on a per-token basis rather than having this scattered throughout several loops over the specs and schema.