Add default value to support web app credentials policies (#2537)
Specify default values for
`Microsoft.Web/sites/{name}/basicPublishingCredentialsPolicies/ftp` and
`Microsoft.Web/sites/{name}/basicPublishingCredentialsPolicies/scm`
which are nested resources under a WebApp.

This PR introduces two new "resources" in explicit versions: `v20190801`
to `v20220901`. This doesn't appear in the default version.

- `WebAppScmAllowed`
- `WebAppFtpAllowed`

When creating a WebApp this defaults to "true". A user can create this
new resource to set "allowed" to "false". Therefore, when we delete this
new resource, we set it back to "true".