Avoid re-populating ignored unset properties during Read (refresh)
- Enable refresh during program test.
- Only remove missing ignored properties from the values to be returned for the inputs - not the outputs otherwise this results in an unhelpful diff.

Because deepCopy utility output is not typed as a map, we have to cast it. If this fails (which should be impossible), we fall back to the original value but also request the user files an issue because this indicates a bug .. but we don't want to panic within the provider as this could leave their program in a bad state without a workaround.