Upgrade (#2842)
Fixes #2839 

The `Migrate.VmUptimeArgs` type is incompatible between 2019-10-01 and
2023-03-15 as the property `HoursPerDay` changed from a number to an
integer, which is a breaking change. To avoid this, we exclude pulling
the new `AssessmentOperation` resource into the default version.

To investigate the issue, I edited the flow so we still return and
output the calculated versions on a failure so we can see what new
resources are being added at what version.