Nightly Upgrade (#2863)
Fixes #2858 

Manual upgrade to latest version. Temporarily remove additional
resources from `HybridContainerService` `2023-11-15-preview` as they
conflict with types from `2022-09-01-preview`. We could consider trying
to upgrade the tracking version to `2023-11-15-preview` once we're sure
it's available in all regions.

The breaking changes are not a result of a change on our side, but a
change upstream.