Remove failing arm2pulumi release job
This is broken because goreleaser tries to call a makefile target that doesn't exist.
We don't need to release this any more because we have the new ARM converter template instead.
The web hosted tool also has been unable to be updated in a long time and should be replaced entirely with the ARM converter plugin.