Fully support the OIDC token configuration (#2887)
This PRs remediates an oversight regarding OIDC configuration.

On GitHub, a token exchange local to GitHub happens first where an
internal GitHub token is exchanged for an OIDC token from GitHub. This
exchange is what's configured via
`oidcRequestToken`/`ARM_OIDC_REQUEST_TOKEN` and
`oidcRequestUrl`/`ARM_OIDC_REQUEST_URL`.

However, on other providers such as Azure and GitLab, a ready-to-use
OIDC token is provided. This is configured via the single config
`oidcToken`/`ARM_OIDC_TOKEN`. But that failed this provider's check for
valid configuration, which this PR fixes.

Resolves #2884 
Resolves #2885
May also fix #2609, to be tested.

Commits:
- Upgrade auth deps as far as possible - further requires migrating to
hashicorp/go-azure-sdk
- Support OIDC tokens directly, refactor and unit test
- Explicit provider argument for OIDC token