Try OIDC config in opposite order: token preferred, like in go-azure-helpers (#2899)
There are different ways to configure OIDC authentication, depending on
which service the program runs on (GitHub, Azure DevOps, etc.):
1. a single OIDC token
2. a "request token" and a URL at which the request token can be
exchanged for an OIDC token.

This PR makes sure that the "OIDC token" method is preferred. It's more
common, and it's also the way go-azure-helpers and Terraform providers
behave.

Implementation note: it's currently allowed to configure both methods,
token and request token+url. Only one of them will be used. On the one
hand, this might be a bit confusing for users. On the other hand, this
is again how TF and go-azure-helpers behave. Users might have a scenario
where they configure both methods and the second acts as a fall-through.