Add selected GET invokes (#2922)
We only create `getFoo` or `listFoo` functions in combination with a
`Foo` resource. We don't turn all GET endpoints to functions because
most aren't very useful.

In some cases, like #2419 to start with, we do however miss a GET
function that's users need. This PR adds the one from that issue and
lays the groundwork for adding more in a well-defined place.

Resolves #2419