Allow removing network rule sets from storage accounts (#2915)
Fixes #2507 

In this issue, the user removed a property from their resource
definition and expected it to get removed or reset in Azure as well.
Azure, however, interpreted the omission of the property on the update
request as "no changes to make here".

This leads to out-of-sync Pulumi and Azure state. Refreshing doesn't
work to fix this because when the updated state comes back, the removed
property is still there, so then the value gets maintained across a
refresh because it treats it like a default value.

The spec does not describe this behavior, so I added a custom exception:
on update of a storage account, if the property was set previously but
now is not, explicitly set it to the default. I copied that default from
an Azure response.

The added test in examples is essentially the user's repro from #2507. After the
second update where the network rules were removed, we assert that they
are back to the default value in outputs.