Allow maintaining nested sub-resources at arbitrary depth (#2950)
This PR is a generalization of #2755, and groundwork for #594. I
recommend reading their descriptions first.

The properties pointing to sub-resources that can be maintained as
stand-alone resources can now be at any level, with `#/type` references
being followed. Traversal methods have been added to
`resources.AzureAPIProperty` that can gather the paths leading to
sub-resource properties (and can also be used for other traversals).

In the provider, a new function `findUnsetSubResourceProperties` uses
the collected sub-resource paths to look up these properties in
API-shaped objects and collects the ones that are not set. In `Read()`,
we remove them from outputs so they are not projected into inputs. In
`Update()`, we use them to determine where we need to insert cloud state
to preserve values not set in our inputs.

In order to be able to unit test code that looks up types, this PR also
introduces an abstraction `resources.TypeLookupFunc` over the previous
direct access to `azureNativeProvider.resourceMap.Types` which is hard
to mock.