Diff keyed arrays (#2968)
Azure's specification [has a custom field called
x-ms-identifiers](https://azure.github.io/autorest/extensions/#x-ms-identifiers)
which can be used to specify a unique key for items in the array. Where
this feature is used, the order of the array might not be maintained by
Azure - the order returned by a GET request might not match the order as
sent in our PUT request. For details see #2781.

This PR changes diffing to handle these keyed arrays more like sets,
comparing elements based on their ids.

It also contains @danielrbradley's PR #2777 to generate a report of
these properties.

In order to be able to customize the array diffing, it was necessary to
copy some pulumi/pulumi diffing code to this repository. Quoting #2781,
> the diff is implemented in two steps:
> 1. Calling the generic resource.PropertyMap.Diff.
> 2. Taking the output of the generic diff and attempt to modify it with
Azure-specific details (identify forceNew properties and calculates
replacements).

We needed to modify step 1 here.

Moving the diffing code here also allows us to integrate and simplify
the Azure-specific post-processing of diffs of step 2. That's not part
of this PR.

Resolves #2781

---------

Co-authored-by: Daniel Bradley <daniel@pulumi.com>