Case-insensitive comparison of Azure ids for more correct diffs (#2993)
This PR aims to resolve #854. Basically, the issue is that the common
prefix of Azure IDs,
`/subscriptions/${sub}/resourcegroups/${rg}/providers/microsoft.${namespace}/`,
is case-insensitive, but we diff it as if it weren't. This leads to
unnecessary updates as well as delete-replaces in keyed arrays where the
Azure id property is a key, like in #854.

The PR introduces a string comparison that, if the string looks like an
Azure id, compares the common prefix case-insensitively, but still
compares the part after the prefix case-sensitively since we cannot know
if names of various resource providers are case-insensitive.

We already have the `applyAzureSpecificDiff` func, but it handles only
top-level properties and wouldn't help with keyed arrays, so I went the
route of applying the id normalization to all strings.

The downside is that might be a bit expensive. I tried to delay the
regex matching where possible.

This PR fixes the repro in #854 for me.