Set ForceNew on referenced types (#2970)
_This PR is on top of #2968._

This PR enables auto re-creation on sub-property changes which resolves
#2782. As that issue points out, the reason this was disabled so far was
to prevent over-eager recreation of resources when unnecessary. With
correct diffing of keyed arrays, this might be fixed. Further testing
needed, therefore this is a draft PR.

The core implementation is a rather small change in properties.go. We
already had the data about referenced types to find properties where
`forceNew` should be set, but avoided doing so. Now we can enable this.
We use a new metadata flag rather than the existing `forceNew` so we can
treat the referenced type properties differently, namely to provide an
escape hatch from the new forced recreation.

Since unwanted recreation of resources is very disruptive, this PR
includes an escape hatch, i.e., a way to opt-out of this change. For
now, it's actually an opt-in to be cautious about breaking user programs
in a minor version. To this end, a new environment variable
`PULUMI_FORCE_NEW_FROM_SUBTYPES` can be set to "true" to enable the
behavior described above. _Question: should this be a proper Pulumi
config setting?_

It works in manual testing for the Managed Environment VNet example from
#2577. Flipping a property on the environment's `vnetConfiguration`
object results in

```
❯ PULUMI_DEBUG_PROVIDERS="azure-native:58241" pulumi up -s dev2
Previewing update (dev2)

     Type                                    Name                   Plan        Info
     pulumi:pulumi:Stack                     azn-2577-vnet-ts-dev2              3 warnings
 +-  └─ azure-native:app:ManagedEnvironment  appEnvironment         replace     [diff: ~vnetConfiguration]

Resources:
    +-1 to replace
    4 unchanged

Do you want to perform this update? details
  pulumi:pulumi:Stack: (same)
    [urn=urn:pulumi:dev2::azn-2577-vnet-ts::pulumi:pulumi:Stack::azn-2577-vnet-ts-dev2]
    ++azure-native:app:ManagedEnvironment: (create-replacement)
        [id=/subscriptions/0282681f-7a9e-424b-80b2-96babd57a8a1/resourceGroups/resourceGroupc34a3e63/providers/Microsoft.App/managedEnvironments/appEnvironment7f502e85]
        [urn=urn:pulumi:dev2::azn-2577-vnet-ts::azure-native:app:ManagedEnvironment::appEnvironment]
        [provider=urn:pulumi:dev2::azn-2577-vnet-ts::pulumi:providers:azure-native::default_2_0_0_alpha_0_dev::b2b85f57-0239-406b-81f5-d1c764b43587]
      ~ vnetConfiguration: {
          ~ internal: false => true
        }
    +-azure-native:app:ManagedEnvironment: (replace)
        [id=/subscriptions/0282681f-7a9e-424b-80b2-96babd57a8a1/resourceGroups/resourceGroupc34a3e63/providers/Microsoft.App/managedEnvironments/appEnvironment7f502e85]
        [urn=urn:pulumi:dev2::azn-2577-vnet-ts::azure-native:app:ManagedEnvironment::appEnvironment]
        [provider=urn:pulumi:dev2::azn-2577-vnet-ts::pulumi:providers:azure-native::default_2_0_0_alpha_0_dev::b2b85f57-0239-406b-81f5-d1c764b43587]
      ~ vnetConfiguration: {
          ~ internal: false => true
        }
    --azure-native:app:ManagedEnvironment: (delete-replaced)
        [id=/subscriptions/0282681f-7a9e-424b-80b2-96babd57a8a1/resourceGroups/resourceGroupc34a3e63/providers/Microsoft.App/managedEnvironments/appEnvironment7f502e85]
        [urn=urn:pulumi:dev2::azn-2577-vnet-ts::azure-native:app:ManagedEnvironment::appEnvironment]
        [provider=urn:pulumi:dev2::azn-2577-vnet-ts::pulumi:providers:azure-native::default_2_0_0_alpha_0_dev::b2b85f57-0239-406b-81f5-d1c764b43587]
```

---------

Co-authored-by: Daniel Bradley <daniel@pulumi.com>