Add new report to track properties that get overwritten during flattening (#3013)
The Azure spec has an annotation
[x-ms-client-flatten](https://github.com/Azure/autorest/blob/main/docs/extensions/readme.md#x-ms-client-flatten)
which instructs the schema generator to flatten the property into the
parent object. However, there are cases where this leads to overwriting
a property, creating incorrect schema and SDKs, when inner and outer
property have the same name. I discovered this problem accidentally.
First step is to quantify, which this report is for.