Don't delete Service Plan when deleting last Web App (#3004)
Resolves #1529 

By default, when deleting the only Web App in a Service Plan, Azure
deletes the Service Plan as well. This doesn't mesh with the Pulumi
resource model since the Service Plan is a separate resource.

The default Azure behavior can be customized with a query parameter
`deleteEmptyServerFarm` on the DELETE request. However, so far we don't
have any support for delete parameters. I saw two ways of approaching
this:
1. Add support for delete parameters to the schema, and pass them
through so that the regular delete operation in the provider does the
right thing.
2. Implement a custom resource for Web App, with only the delete method,
re-using as much provider code as possible.

The problem with (1) is that it would expose the `deleteEmptyServerFarm`
parameter via the schema but we'd still not know which value it would be
set to without custom handling for the Web App case. This PR implements
approach (2).

In order to be able to re-use the provider's existing delete logic, the
required methods are encapsulated in a new interface `AzureClient` which
can be passed to the custom resources factory. This will probably also
be useful to support POST-based operations, as it would be simple to add
the `azurePost` method to the interface and call it from custom
resources.