Don't add empty objects that may be of the wrong type when maintaining subresource properties (#3040)
Issue #3036 happens because of the logic to maintain unset sub-resource
properties.

In the Update path, in `maintainSubResourcePropertiesIfNotSet` ->
`writePropertiesToBody`, the property to maintain is `properties.
privateEndpointConnections`. The remote state doesn't have it, either.
So the code adds an empty container `privateEndpointConnections:
map[string]any` to the body, then the algorithm stops as there's no more
remote state to traverse.

However, `privateEndpointConnections: {}` is wrong because this property
is of type array, and should be `privateEndpointConnections: []`.

This PR removes the addition of empty containers when there's no
corresponding remote state. This does change existing logic, as one
changed unit test shows. As far as I can see, the new algorithm makes
more sense: leaving out the empty container matches the remote state.

Resolves #3036