Enable RequireEmptyPreviewAfterRefresh in all tests (#3056)
This PR enables
[`RequireEmptyPreviewAfterRefresh`](https://github.com/pulumi/pulumi/pull/15309)
for all tests except three that fail with it. There is no functional
provider change here. A follow-up PR with fix the provider for those
three tests and re-enable them.

I had to sprinkle a few `ignoreChanges` in several places. Once
https://github.com/pulumi/pulumi-azure-native/issues/3049 is fixed, we
can remove some of them. The remaining ones are legit, and I added
comments why.