Reset, don't remove sub-resource properties to avoid wrong diffs (#3054)
This PR fixes a defect introduced by #2950. When handling sub-resource
properties that may be maintained as stand-alone resources, we need to
overwrite them in several code paths. Before, we set them to their
default value `[]` on Create, but removed them entirely on Read. This
causes unnecessary diffs on Refresh. This PR standardizes the handling
on resetting to the default value `[]`.

Resolves #3049