Distinguish nil from empty map as default state in resource discovery (#3076)
This PR fixes what I believe is a bug introduced in #2456. For the
default state of a resource, we need to distinguish between `nil` (no
default) and the empty object as the empty default. This is also
described in the doc comment of `AzureAPIResource.DefaultBody` in
resources.go:
> - omitempty is not set to distinct between nil (no default) and empty
maps (empty default)

`AzureClient.CanCreate` checks `DefaultBody != nil` to detect resources
that are automatically created with their parent.

The actual change in metadata is not visible in the PR because we don't
commit it. It looks like this:
```diff
--- ../m-master.json    2024-02-12 20:22:54.578408624 +0100
+++ ../m.json   2024-02-12 20:21:14.881957652 +0100
@@ -1005922,7 +1005922,7 @@
           ]
         }
       },
-      "defaultBody": {},
+      "defaultBody": null,
       "putAsyncStyle": "azure-async-operation",
       "deleteAsyncStyle": "azure-async-operation"
     },
```