Reorganize examples integration tests into long and short based on run times (#3095)
This PR was motivated by #3078 because this new test has a run time of
~24m, too long for PRs. That motivated me to collect data on the run
times of all our e2e tests. Until now, about half of the tests were
classified as "long", but with no clear pattern.

The data shows that two thirds of the tests take 7 minutes or less, so
we picked that as a cut-off. Those tests are now classified as "short".
PRs will now only run the short tests, leading to faster builds with
still two thirds of the previous test coverage.