Use v4 consitently for upload/download-artifact (#3111)
the `download-artifact@v3` action can't pull artifacts uploaded by
`upload-artifact@v4` (and vice versa). It looks like we updated some but
not all of our builds to use v4 of these actions, which is causing
master builds to fail. This change moves everything to v4 which should
fix the builds.

Fixes: https://github.com/pulumi/pulumi-azure-native/issues/3109