Add invokes that return non-object types such as plain strings
If the schema has no properties but is a primitive output, we include it as a property so invokes can be generated.
Ideally we'd just represent it as a primitive type and use FunctionSpec.ReturnType to specify this, but pulumi/pulumi
15739 and 15738 prevent this.

Unfortunately, this required adding an unfortunate third boolean param
to genProperties to limit this new case to function responses only.