Add support for Palo Alto Networks (#3182)
Resolves #2845 

Palo Alto Networks is an independent company offering certain resources
on Azure as an ISV - see [Secure your cloud environment with Cloud
Next-Generation Firewall by Palo Alto Networks, an Azure Native ISV
Service](https://azure.microsoft.com/en-us/blog/secure-your-cloud-environment-with-cloud-nextgeneration-firewall-by-palo-alto-networks-an-azure-native-isv-service-now-in-preview/).

We weren't including these resources because in the OpenAPI spec, they
are not under the _"Microsoft.Service"_ prefix, but the
_"PaloAltoNetworks.Service"_ prefix. Other than that, they are standard
OpenAPI specs, so this PR simply adds this prefix.

Something to consider is that we simply drop the prefix in later
processing. So whether a service is Microsoft.Foo or
PaloAltoNetworks.Foo, it simply becomes _Foo_. Currently, there are no
naming clashes, but theoretically there could be.