Enable generateExtraInputType for Go
GenerateExtraInputTypes makes the code generator generate input and
output types for all plain types, instead of for only types that are
used as input/output types.

We need this here because the Go SDK is special in that it's generated
as separate modules, from a sub-schema per module. This causes SDK-gen
to not infer the type used from a union as input.

Unfortunately, this change increases the Go SDK size by about 19%.

Addresses #1554