Detect duplicate key values when diffing keyed arrays (#3320)
The Azure spec has "keyed arrays" which are essentially sets. This PR
resolves #3311 which showed that some keyed arrays actually have
duplicate key values, which defeats the purpose and is invalid (filed
Azure/azure-rest-api-specs#29288 upstream).

This PR adds detection of this degenerate case, upon which we abort the
keyed array diff and continue with a regular array diff.


Tested locally with the exact repro of #3311.