Allow keyvault accesspolicy to have object id in its id (#3295)
Fixes https://github.com/pulumi/pulumi-azure-native/issues/3282

For import to work, we need to be able to identify which access policy
we're addressing and therefore need the object id to be part of the id.

We're unable to update the `id` of existing resources with the current
provider implementation structure so need to be able to continue to
parse ids in both the old and new format. It might be possible to
refactor the provider to allow custom resources to return a new `id` but
this should be undertaken in as follow-up work.

The Path property of the custom resource is used as an internal
identifier, but this is looked up from the metadata which is generated
by the resource type which is not changing, so this is backwards
compatible.

The `crudClient.PrepareAzureRESTIdAndQuery` is that part which builds
the `id`. `objectId` doesn't exist as a top-level property so we set it
from within the `policy` instead (`policy.objectId`) - for which we also
have to allow paths to contain references to nested properties.

## Testing

The added keyvault-accesspolicy test-program can be used to test the two
scenarios for this change: id compatibility and import.

### 1) Upgrading

1. Deploy the program with a previous version of the provider (e.g.
2.43.0)
2. Upgrade the provider to include this change (e.g. via attaching to
the running dev version)
3. Perform another `up` and `refresh` to ensure no changes are made.

Note: this can be targeted using `pulumi refresh -t
urn:pulumi:dev::keyvault-accesspolicy::azure-native:keyvault:AccessPolicy::ap1`

### 2) Importing

Using the new version of the provider, import the access policy back in
with a different name (e.g. `ap2`):

```
pulumi import azure-native:keyvault:AccessPolicy ap2 "$(pulumi stack output ap1Id)/$(pulumi stack output ap1ObjectId)"
```

### Automation

Ideally, these tests would be run via automation. However, there's no
current test infrastructure written for handling upgrade testing or
import testing so this will have to be done as follow-up work.