Support OIDC file token authentication (#3382)
One typical way of authenticating via OIDC is that the identity provider
supplies the OIDC token in a file. This provider already had support for
this built in to the auth stack, but due to an oversight, validation
wouldn't consider it and the provider would error before attempting to
authenticate. This PR fixed this oversight and resolves #3373.

pulumi/registry#4790 updates the docs accordingly.