Don't require blob access tier as output and don't store it when empty (#3425)
Fixes #3414

The user got the error `ValueError: '' is not a valid BlobAccessTier`
when creating a storage blob. The reason is that the provider always
stored the access tier returned from Azure in the output. However, for
premium storage accounts, the access tier is implicitly 'premium' and
not returned. That caused an empty string to be stored in Pulumi state
which is not a valid instance of the access tier enum.

[Blob Storage feature support in Azure Storage
accounts](https://learn.microsoft.com/en-us/azure/storage/blobs/access-tiers-overview)
details where access tiers are supported.

Tested locally. The user's repro program from #3414 works with this fix.